+++ 17b.c
@@ -4,7 +4,7 @@ typedef struct {
     int64_t a;
     int64_t b;
     int64_t c;
-    uint16_t pc;
+    uint8_t pc;
 } Computer;
 
 #define combo(o) ((o) < 4 ? (o) : (o) == 4 ? c.a : (o) == 5 ? c.b : c.c)
